requirements.txt
fastapi>=0.90.0
uvicorn==0.22.0
streamlit>=1.18.0
pandas>=2.0.0
numpy>=1.25.0
google-cloud-storage>=2.11.0
google-cloud-dlp>=3.14.0
vertexai>=1.11.0
tenacity>=8.2.0
requests>=2.31.0
python-multipart>=0.0.6
plotly>=5.18.0
openai>=1.0.0
pinecone-client>=2.2.1